requirement.txt
django~=3.9.2
pillow
django-debug-toolbar
django-extensions
django-bootstrap4